feat: Add create_category function to create a new category
This commit adds a new function `create_category` to the `create_category.py` file in the `src/core/category/application` directory. The function takes in parameters for `name`, `description`, and `is_active` and creates a new `Category` instance with the provided values. The function returns the `id` of the created category. This function will be used to create new categories in the application.